deprecating bok-choy tasks from rake to paver

+++ pavelib/utils/process.py
@@ -21,7 +21,7 @@ def kill_process(proc):
         os.kill(child_pid.pid, signal.SIGKILL)
 
 
-def run_multi_processes(cmd_list, out_log=None, err_log=None):
+def run_multi_processes(cmd_list, out_log=None, err_log=None, background=False, cwd=None):
     """
     Run each shell command in `cmd_list` in a separate process,
     piping stdout to `out_log` (a path) and stderr to `err_log` (also a path).
@@ -29,7 +29,7 @@ def run_multi_processes(cmd_list, out_log=None, err_log=None):
     Terminates the processes on CTRL-C and ensures the processes are killed
     if an error occurs.
     """
-    kwargs = {'shell': True, 'cwd': None}
+    kwargs = {'shell': True, 'cwd': cwd}
     pids = []
 
     if out_log:
@@ -44,27 +44,42 @@ def run_multi_processes(cmd_list, out_log=None, err_log=None):
         for cmd in cmd_list:
             pids.extend([subprocess.Popen(cmd, **kwargs)])
 
-        def _signal_handler(*args):
-            print("\nEnding...")
+        if not background:
+            def _signal_handler():
+                """
+                What to do when process is ended
+                """
+                print("\nEnding...")
 
-        signal.signal(signal.SIGINT, _signal_handler)
-        print("Enter CTL-C to end")
-        signal.pause()
-        print("Processes ending")
+            signal.signal(signal.SIGINT, _signal_handler)
+            print("Enter CTL-C to end")
+            signal.pause()
+            print("Processes ending")
 
-    except Exception as err:
+    except Exception as err:  # pylint: disable-msg=broad-except
         print("Error running process {}".format(err), file=sys.stderr)
 
     finally:
-        for pid in pids:
-            kill_process(pid)
+        if not background:
+            for pid in pids:
+                kill_process(pid)
+
+    return pids
 
 
-def run_process(cmd, out_log=None, err_log=None):
+def run_process(cmd, out_log=None, err_log=None, cwd=None):
     """
     Run the shell command `cmd` in a separate process,
     piping stdout to `out_log` (a path) and stderr to `err_log` (also a path).
 
     Terminates the process on CTRL-C or if an error occurs.
     """
-    return run_multi_processes([cmd], out_log=out_log, err_log=err_log)
+    return run_multi_processes([cmd], out_log=out_log, err_log=err_log, cwd=cwd)
+
+
+def run_background_process(cmd, out_log=None, err_log=None, cwd=None):
+    """
+    Runs a command as a background process. Note you will have to kill the processes
+    explicitly when you are done with them, so the pids are returned.
+    """
+    return run_multi_processes([cmd], out_log=out_log, err_log=err_log, background=True, cwd=cwd)
